Missing data also drops from random effects
