+++ bambi/models.py
@@ -4,6 +4,7 @@ import pymc3 as pm
 from pymc3.model import FreeRV
 import matplotlib.pyplot as plt
 from bambi.external.six import string_types
+from bambi.external.patsy import Ignore_NA
 from collections import OrderedDict, defaultdict
 from bambi.utils import listify
 from patsy import dmatrices, dmatrix
@@ -124,8 +125,14 @@ class Model(object):
                 % na_index.sum()
             warnings.warn(msg)
             keeps = np.invert(na_index)
-            for t in self.fixed_terms.values():
-                t.data = t.data[keeps]
+            for t in self.terms.values():
+                # remove missing values from random effects
+                if isinstance(t.data, dict):
+                    t.data[list(t.data.keys())[0]] = \
+                        t.data[list(t.data.keys())[0]][keeps]
+                # remove missing values from fixed effects
+                else:
+                    t.data = t.data[keeps]
             self.y.data = self.y.data[keeps]
 
         # X = fixed effects design matrix (excluding intercept/constant term)
@@ -657,34 +664,4 @@ class Term(object):
             self.levels = list(range(data.shape[1]))
 
         self.data = data
-
-class Ignore_NA(object):
-    """
-    Custom patsy.missing.NAAction class to force Patsy to ignore missing values.
-    See Patsy code/API for NAAction documentation.
-    """
-    def __init__(self, on_NA="ignore", NA_types=["None", "NaN"]):
-        self.on_NA = on_NA
-        if isinstance(NA_types, str):
-            raise ValueError("NA_types should be a list of strings")
-        self.NA_types = tuple(NA_types)
-
-    def is_categorical_NA(self, obj):
-        if "NaN" in self.NA_types and safe_scalar_isnan(obj):
-            return True
-        if "None" in self.NA_types and obj is None:
-            return True
-        return False
-
-    def is_numerical_NA(self, arr):
-        mask = np.zeros(arr.shape, dtype=bool)
-        if "NaN" in self.NA_types:
-            mask |= np.isnan(arr)
-        if mask.ndim > 1:
-            mask = np.any(mask, axis=1)
-        return mask
-
-    def handle_NA(self, values, is_NAs, origins):
-        return values
-
-
+
